db: structure: wire an EventStore to DB entities
The motivation for introducing an EventStore is two-fold:

1. To support external stores for build events to avoid filling the core
   database with build logs. For instance, we could add an ElasticSearch
   EventStore for managing build events, while keeping everything else
   in Postgres
2. To support arbitrary Event processing before storing to the final
   EventStore. For instance, secret redaction is currently implemented
   by having a modified `dbEventWriterWithSecretRedaction` that
   essentially just delegates to the main `dbEventWriter` after
   redacting any secrets. This could possibly be implemented as an
   `EventStore` following the Decorator pattern that receives full log
   events, redacts secrets from them, and then delegates to another
   EventStore. While the initial approach works, it's difficult to add
   more event processing - the initial motivation for this was #3879.
   Another thing this lets us do is e.g. syslog forwarding in real time,
   rather than having a syslog drainer component.

This takes an alternative approach to my initial attempted refactor that
can began in this commit: 8eff6c7e0a81e7ac6baed9dbe571be4e3f0c3a40

The main difference is that everything is contained in the `db` package
here, rather than passing the `EventStore` to every component that could
perform an action that produces events. Additionally, I chose to use the
name `EventStore` instead of `EventProcessor`. While I argued that a
generic name like `EventProcessor` would be beneficial to encourage
general data transformations, I feel like it's a bit too general - and,
if we follow the Decorator pattern, every `EventProcessor` will
necessarily need to be an `EventStore` as well, so might as well just
call it what it is.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>