atc: behaviour: spike on Elasticsearch EventStore
At this point, we have an implementation of the EventStore that
communicates with Elasticsearch and seems to pretty much work.
This commit doesn't address the configuration for Elasticsearch -
currently, it doesn't use any auth, and atccmd is just hardcoded to
point to the docker-compose'd Elasticsearch.

It also only works with Elasticsearch 7 at the moment, though with a bit
of extra work it could be adjusted to work with 6. The implementation
makes use of ILM (index lifecycle management) which is a really cool
feature of Elasticsearch. Basically, you can provide an ILM policy which
specifies different index "phases" based on how stale the index is (hot,
warm, cold, and delete), and you can perform different actions depending
on the phase. We specify that "hot" indexes should automatically rollover
after they reach 50GB or every 30 days (whichever comes first). 7 days
after they rollover, they'll move to "warm", where data is read-only and
the storage gets optimized to take up less disk space. 30 days after
rollover, indexes move to "cold" where they are frozen - nodes don't
allocate memory for them, and the data isn't included in searches by
default, but the data can still be fetched if it's needed. I didn't make
use of it yet, but we can also specify node classes depending on the
phase (so, you could reserve the high I/O nodes for "hot" indexes, and
have cheaper/slower nodes for "warm" and "cold" indexes). I pretty much
followed this guide on this stuff: https://www.elastic.co/blog/implementing-hot-warm-cold-in-elasticsearch-with-index-lifecycle-management

Since Elasticsearch doesn't support auto-incremented ids (unlike
Postgres), data is keyed as follows: first by the event timestamp, and
if multiple events have the same timestamp, the tie is broken by an
in-memory counter. Since in theory there could be multiple ATCs saving
events for the same build, each ATC's counter starts at a random count
to avoid the case where two events have the same timestamp and the same
counter value.

One thing I learned that isn't ideal is that Elasticsearch indeces only
refresh with newly added documents every 1 second by default. What that
means is if we just insert documents normally and then call `Get`
immediately after (which we do in the EventSource), those newly added
documents won't be there by default. The way I got around this was to
add `Refresh("wait_until")` to the Bulk insert call, which will cause
the request to block until the index has been refreshed, which could be
up to a second. Another option would be to call `Refresh("true")` which
force-refreshes the index, but that isn't recommended. I have an idea
that involves extending our use of the Postgres notification bus to get
around this in the general case, which I'll explore next.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>