web: structure: revert Dashboard.pipelines to FetchResult
This commit reverts part of d196102f510f0cd4c2e58690a81507e75d4bbe31
which changed `Dashboard.Model.pipelines` to a `Maybe` instead of a
`FetchResult`. While we can still keep the `.stale` field on the
pipeline records, it's useful to know if the pipelines we're displaying
came from the cache.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>